secure: Buffer overflow in amstream driver (due to integer overflow) [1/1]
PD#OTT-5057

Problem:
When convert 64bit to 32bit, truncation happens

Solution:
avoid the truncation

Verify:
u212

+++ drivers/stream_input/parser/thread_rw.c
@@ -151,7 +151,7 @@ static ssize_t threadrw_write_in(
 {
 	int ret = 0;
 	int off = 0;
-	int left = count;
+	size_t left = count;
 	int wait_num = 0;
 	unsigned long flags;
